chore: update toolchain versions (#439)
Since rust-analyzer has been merged into the main rustlang repo, it is
tested for breaking changes. This means that we don't need to pin the
nightly to a specific version anymore, as using the latest nightly is
always guaranteed to work with the latest version of the rust-analyzer
extension (try it yourselves, for me it finally just works).

This PR then does the following:

- update the toolchain file to use the latest nightly, which gives us
access to potentially new clippy warnings, and to better rust-analyzer
performances
- update the runtime builder to use Rust 1.64 as also nowadays done by
Polkadot
- remove the deprecated `cargo-features = ["workspace-inheritance"]`,
since from 1.64 it has been stabilised (this could be caught only by
using a recent enough nightly version)
- Let the CI job use the same version as in the toolchain file, by
removing the `+nightly` flag. This will ensure we have a consistent
behaviour between local machine and CI server.

The only thing to consider is that the nightly will differ from the
version that srtool uses to build the runtime WASM. That was the case
also before, and it will always be until substrate will support
compilation with the stable toolchain.