feat: genesis config for linking pallet (#476)
## No Ticket

* adds accounts for doc tests to endowed accounts
* Linking Pallet
* MigrationState is Done in genesis (fix
https://github.com/KILTprotocol/ticket/issues/2475)
    * allow to configure pallet in genesis


I added the option to configure the DidLookupPallet in genesis, so that
we can start a test chain with already populated DID-Account links. We
could start adding genesis configs to the other pallets as well. Some
tests require well known DIDs or w3n to exist. We could add them in the
genesis config, so that the tests work out of the box with dev chains.