feat: bump number of URLs for a service endpoint to 2 (#517)
Fixes https://github.com/KILTprotocol/ticket/issues/2633.

It also removes some unused constants and adjusts the value of the max
encoded length for both a service endpoint and a DID key, keeping in
mind the additional bytes that a `BoundedVec` stores along with the
actual data. The padding for a service endpoint is added after running
the test, since the number of additional bytes depends on the encoding
of the maximum limit for it. Tests as they are run now should ALWAYS
catch any changes in this value.