fix: potential circular dependency (#547)
The lookup pallet introduced a dependency on the `runtime-common` crate,
which should not really be there. While working on DIP, where the
reverse relationship is introduced, we have a circular dependency.

This PR abstracts in the lookup pallet over the definition of `Balance`
by re-using the trait constraints that the `pallet-balances` defines. It
also resolves one lint warning regarding an unused function.